# Improve performance for applications with many layers - Backend Part