--- a/src/Mapbender/CoreBundle/Resources/config/services.xml
+++ b/src/Mapbender/CoreBundle/Resources/config/services.xml
@@ -199,6 +199,7 @@
             <argument type="service" id="event_dispatcher" />
             <argument type="service" id="mapbender.element_filter" />
             <argument type="service" id="mapbender.source.typedirectory.service" />
+            <argument type="service" id="Doctrine\ORM\EntityManagerInterface" />
             <argument type="service" id="mapbender.source.url_processor.service" />
             <argument type="service" id="router" />
             <argument type="service" id="assets._default_package" />