--- a/src/Mapbender/WmsBundle/Resources/config/services.xml
+++ b/src/Mapbender/WmsBundle/Resources/config/services.xml
@@ -36,6 +36,7 @@
         <service id="mapbender.source.wms.config_generator" class="%mapbender.source.wms.config_generator.class%">
             <argument type="service" id="mapbender.source.url_processor.service" />
             <argument type="service" id="security.token_storage" />
+            <argument type="service" id="doctrine.orm.default_entity_manager" />
             <argument>%wms.default_layer_order%</argument>
         </service>
         <service id="mapbender.source.wms.instance_factory" class="%mapbender.source.wms.instance_factory.class%">