This is a follow-up to #1744.

Having many layers in an application significantly slowed down loading time, on the one hand in the frontend when generating the layer tree, but also in the backend when creating the application config.

It was discovered, that for resolving sublayers, doctrine makes a separate database call for every single instance layer. Also, using the xdebug profiling tool, it was discovered that doctrine's hydration of the individual database objects takes a lot of time (~30% of the whole processing time), specifically hydrating WmsInstanceLayers. 

As a consequence, the WmsSourceInstanceConfigGenerator was refactored to avoid hydrating objects for each layer but instead have a single database call returning an array with the data. 
The downside is that some code needs to be duplicated and the code is less readable and less object-oriented. However, the performance boost is huge. 

The table illustrates the performance metrics in an exemplary application having 944 instance layers in the application. Each time, the application config was called 100 times.

|                           | **before refactoring**           | **after refactoring**           |
|---------------------------|:-------------------------------:|:-------------------------------:|
| **Individual database calls** | 1331                          | 216                             |
| **Min**                   | 1133.146 ms                     | 510.181 ms                      |
| **Max**                   | 1634.287 ms                     | 833.666 ms                      |
| **Median**                | 1225.509 ms                     | 561.619 ms                      |
| **Stddev**                | 78.099 ms                       | 38.960 ms                       |